Initial commit, introduce the idea of an automatic backuper
The idea started off as a WordPress-only backuper, but I'm just going to
make this a file and database backuper. We already have the regex to get
the database infos from `wp-config.php`, though... So I will likely keep
that and put a setting in `config.yaml` like isWordPress (boolean).

Signed-off-by: Dave Russell Jr <dave@createazure.com>