Introduce the configuration file in the pl file
This commit will restructure our code a bit to introduce a YAML
configuration file. This is simply so that we can declare default
settings without editing the perl script itself.

This also introduces the error logging subroutine, so that we can log to
a file if we want, because STDOUT when cronjob is sloppy.

Signed-off-by: Dave Russell Jr <dave@createazure.com>